requirements.txt
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
Pillow==11.2.1
plotly==6.1.1
scikit_learn==1.6.1
seaborn==0.13.2
streamlit==1.45.1
